rename imagecapturer
